# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Mini E-Commerce API
  version: 1.0.0
  description: API for user registration, login, product management, orders, and cart operations.
  contact:
    name: TurynX
    email: your@email.com
    url: https://your-site.com
servers:
  - url: http://localhost:3333
...
security:
  - bearerAuth: []
paths:
  /users/register:
    post:
      operationId: registerUser
      summary: Register a new user
      description: Creates a new user account.
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        "201":
          description: User successfully created
        "400":
          description: Invalid input
...
  /products/{id}:
    put:
      operationId: updateProduct
      summary: Update a product
      description: Updates product details.
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                  minimum: 0
      responses:
        "200":
          description: Product updated
        "404":
          description: Product not found
    delete:
      operationId: deleteProduct
      summary: Delete a product
      description: Deletes a product by ID.
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
...
  /cart/remove/product/{id}:
    delete:
      operationId: removeFromCart
      summary: Remove product from cart
      description: Removes a product from the user's cart.
      tags: [Cart]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product removed
        "404":
          description: Product not found in cart
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT